Added comment of how to access token from console.
A key part about this project is trying to move it to a format that
people can use without SW Developer knowledge. Improving the "token
scraping" from manually accessing Response headers looking for the value
to a simple console command is a big step towards this.

The README.md file has been improved to help improve legibility.